Add types to more closely reflect the API
