+++ Octokit/Models/Request/NewTag.cs
@@ -13,7 +13,7 @@ namespace Octokit
         public string Object { get; set; }
         [SuppressMessage("Microsoft.Naming", "CA1721:PropertyNamesShouldNotMatchGetMethods", Justification = "Property name as defined by web api")]
         public TaggedType Type { get; set; }
-        public Signature Tagger { get; set; }
+        public SignatureResponse Tagger { get; set; }
         internal string DebuggerDisplay
         {
             get
