+++ Octokit/Models/Response/Commit.cs
@@ -7,8 +7,8 @@ namespace Octokit
     public class Commit : GitReference
     {
         public string Message { get; set; }
-        public Signature Author { get; set; }
-        public Signature Committer { get; set; }
+        public SignatureResponse Author { get; set; }
+        public SignatureResponse Committer { get; set; }
         public GitReference Tree { get; set; }
         public IEnumerable<GitReference> Parents { get; set; }
         public int CommentCount { get; set; }
